allow also SACD iso images with line ending as ".sacd"
This is added once to allow users to use *.iso as an extension instead.

In addition, there are currently problems in Kodi that "* .iso" is handled incorrectly in its media scan and fails when this add-on is activated.
This change does not fix it, but the user has the option of removing ".iso" from the addon.xml and renaming his files to ".sacd" as the extension.